backend/procurement-explorer/requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.4.0; python_version >= '3.8'
aiohttp==3.10.5; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
alembic==1.13.2; python_version >= '3.8'
altair==5.4.0; python_version >= '3.8'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.4.0; python_version >= '3.8'
asgiref==3.8.1; python_version >= '3.8'
async-timeout==4.0.3; python_version < '3.11'
attrs==24.2.0; python_version >= '3.7'
azure-core==1.30.2; python_version >= '3.8'
azure-identity==1.17.1; python_version >= '3.8'
azure-storage-blob==12.22.0; python_version >= '3.8'
backoff==2.2.1; python_version >= '3.7' and python_version < '4.0'
bcrypt==4.2.0; python_version >= '3.7'
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
blinker==1.8.2; python_version >= '3.8'
bs4==0.0.2
build==1.2.1; python_version >= '3.8'
cachetools==5.5.0; python_version >= '3.7'
certifi==2024.7.4; python_version >= '3.6'
cffi==1.17.0; platform_python_implementation != 'PyPy'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
chroma-hnswlib==0.7.6
chromadb==0.5.5; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
coloredlogs==15.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cryptography==43.0.0; python_version >= '3.7'
dataclasses-json==0.6.7; python_version >= '3.7' and python_version < '4.0'
deprecated==1.2.14; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
docx2txt==0.8
exceptiongroup==1.2.2; python_version < '3.11'
fastapi==0.112.1; python_version >= '3.8'
filelock==3.15.4; python_version >= '3.8'
flake8==7.1.1; python_full_version >= '3.8.1'
flatbuffers==24.3.25
frozenlist==1.4.1; python_version >= '3.8'
fsspec==2024.6.1; python_version >= '3.8'
gitdb==4.0.11; python_version >= '3.7'
gitpython==3.1.43; python_version >= '3.7'
google-auth==2.34.0; python_version >= '3.7'
googleapis-common-protos==1.63.2; python_version >= '3.7'
greenlet==3.0.3; python_version < '3.13' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
grpcio==1.65.5; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
html2text==2024.2.26; python_version >= '3.8'
httpcore==1.0.5; python_version >= '3.8'
httptools==0.6.1
httpx==0.27.0; python_version >= '3.8'
huggingface-hub==0.24.6; python_full_version >= '3.8.0'
humanfriendly==10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.7; python_version >= '3.5'
importlib-metadata==8.0.0; python_version >= '3.8'
importlib-resources==6.4.4; python_version >= '3.8'
isodate==0.6.1
jinja2==3.1.4; python_version >= '3.7'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
kubernetes==30.1.0; python_version >= '3.6'
langchain==0.3.1; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-community==0.3.1; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-core==0.3.7; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-text-splitters==0.3.0; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-openai==0.2.1; python_version < '4.0' and python_full_version >= '3.8.1'
langsmith==0.1.129; python_version < '4.0' and python_full_version >= '3.8.1'
lxml==5.3.0; python_version >= '3.6'
mako==1.3.5; python_version >= '3.8'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
marshmallow==3.22.0; python_version >= '3.8'
mccabe==0.7.0; python_version >= '3.6'
mdurl==0.1.2; python_version >= '3.7'
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msal==1.30.0; python_version >= '3.7'
msal-extensions==1.2.0; python_version >= '3.7'
multidict==6.0.5; python_version >= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
narwhals==1.5.2; python_version >= '3.8'
numpy==1.26.4; python_version >= '3.9'
oauthlib==3.2.2; python_version >= '3.6'
onnxruntime==1.19.0
openai==1.51.0; python_version >= '3.8'
opentelemetry-api==1.26.0; python_version >= '3.8'
opentelemetry-exporter-otlp-proto-common==1.26.0; python_version >= '3.8'
opentelemetry-exporter-otlp-proto-grpc==1.26.0; python_version >= '3.8'
opentelemetry-instrumentation==0.47b0; python_version >= '3.8'
opentelemetry-instrumentation-asgi==0.47b0; python_version >= '3.8'
opentelemetry-instrumentation-fastapi==0.47b0; python_version >= '3.8'
opentelemetry-proto==1.26.0; python_version >= '3.8'
opentelemetry-sdk==1.26.0; python_version >= '3.8'
opentelemetry-semantic-conventions==0.47b0; python_version >= '3.8'
opentelemetry-util-http==0.47b0; python_version >= '3.8'
orjson==3.10.7; python_version >= '3.8'
overrides==7.7.0; python_version >= '3.6'
packaging==24.1; python_version >= '3.8'
pandas==2.2.2; python_version >= '3.9'
pillow==10.4.0; python_version >= '3.8'
portalocker==2.10.1; python_version >= '3.8'
posthog==3.5.2
protobuf==4.25.4; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
pyarrow==17.0.0; python_version >= '3.8'
pyasn1==0.6.0; python_version >= '3.8'
pyasn1-modules==0.4.0; python_version >= '3.8'
pycodestyle==2.12.1; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pydantic==2.8.2; python_version >= '3.8'
pydantic-core==2.20.1; python_version >= '3.8'
pydantic-settings==2.4.0; python_version >= '3.8'
pydeck==0.9.1; python_version >= '3.8'
pyflakes==3.2.0; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
pyjwt[crypto]==2.9.0; python_version >= '3.8'
pypdf==4.3.1; python_version >= '3.6'
pypika==0.48.9
pyproject-hooks==1.1.0; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-docx==1.1.2; python_version >= '3.7'
python-dotenv==1.0.1; python_version >= '3.8'
python-multipart==0.0.9; python_version >= '3.8'
pytz==2024.1
pyyaml==6.0.2; python_version >= '3.8'
referencing==0.35.1; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
rich==13.7.1; python_full_version >= '3.7.0'
rpds-py==0.20.0; python_version >= '3.8'
rsa==4.9; python_version >= '3.6' and python_version < '4'
setuptools==73.0.1; python_version >= '3.8'
shellingham==1.5.4; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.1; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
soupsieve==2.6; python_version >= '3.8'
sqlalchemy==2.0.32; python_version >= '3.7'
starlette==0.38.2; python_version >= '3.8'
streamlit==1.37.1; python_version >= '3.8' and python_full_version != '3.9.7'
sympy==1.13.2; python_version >= '3.8'
tenacity==8.5.0; python_version >= '3.8'
tokenizers==0.20.0; python_version >= '3.7'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomli==2.0.1; python_version < '3.11'
tornado==6.4.1; python_version >= '3.8'
tqdm==4.66.5; python_version >= '3.7'
typer==0.12.4; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
typing-inspect==0.9.0
tzdata==2024.1; python_version >= '2'
urllib3==2.2.2; python_version >= '3.8'
uvicorn[standard]==0.30.6; python_version >= '3.8'
#uvloop==0.20.0
watchdog==4.0.2; platform_system != 'Darwin'
watchfiles==0.23.0
websocket-client==1.8.0; python_version >= '3.8'
websockets==13.0
wrapt==1.16.0; python_version >= '3.6'
yarl==1.9.4; python_version >= '3.7'
zipp==3.20.0; python_version >= '3.8'
pymongo==4.8.0
langchain-community
